+++ src/api/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to
   performances)
 - Upgrade fastapi to `0.112.2`
 - Upgrade typer to `0.12.5`
+- Upgrade sqlmodel to `0.0.22`
 
 ## [0.11.0] - 2024-08-14