+++ src/api/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.12.0] - 2024-09-05
+
 ### Added
 
 - Implement GZip requests support (mostly for bulk endpoints)
@@ -175,7 +177,8 @@ and this project adheres to
 
 - Implement base FastAPI app
 
-[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.11.0...main
+[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.12.0...main
+[0.12.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.8.0...v0.9.0